.history/requirements_20250831134743.txt
# AI Security Demo - Banking Assistant
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# AI & LLM
openai==1.3.7
langchain==0.0.340
langchain-openai==0.0.2
chromadb==0.4.18

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data & Utilities
pydantic==2.5.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3

# Testing & Monitoring
pytest==7.4.3
httpx==0.25.2